requirements.txt
Flask
Flask-SocketIO
python-dotenv
gunicorn

